+++ frontend/app/src/components/profitloss/ReportActionable.vue
@@ -43,14 +43,18 @@ const { t } = useI18n();
   <div v-if="actionableItemsLength" class="flex">
     <VDialog v-model="mainDialogOpen" max-width="1000">
       <template #activator="{ on }">
-        <VBtn color="error" depressed v-on="on">
+        <RuiButton color="error" v-on="on">
           <span class="pr-2">
             {{ t('profit_loss_report.actionable.show_issues') }}
           </span>
-          <VChip x-small class="px-2" color="error darken-2">
-            {{ actionableItemsLength }}
-          </VChip>
-        </VBtn>
+          <template #append>
+            <RuiChip
+              size="sm"
+              class="!text-white !p-0"
+              :label="actionableItemsLength"
+            />
+          </template>
+        </RuiButton>
       </template>
       <ReportActionableCard
         v-if="mainDialogOpen"