+++ frontend/app/src/utils/download.ts
@@ -1,3 +1,5 @@
+import { type AxiosResponse } from 'axios';
+
 export const downloadFileByUrl = (url: string, fileName: string): void => {
   const link = document.createElement('a');
   link.setAttribute('href', url);
@@ -9,4 +11,24 @@ export const downloadFileByUrl = (url: string, fileName: string): void => {
   } else {
     link.click();
   }
+};
+
+export const downloadFileByBlobResponse = (
+  response: AxiosResponse,
+  filename: string
+) => {
+  const url = window.URL.createObjectURL(response.request.response);
+  downloadFileByUrl(url, filename);
+};
+
+export const downloadFileByTextContent = (
+  text: string,
+  filename: string,
+  type: 'plain' | 'json' | 'csv' = 'plain'
+) => {
+  const file = new Blob([text], {
+    type: `text/${type}`
+  });
+  const url = window.URL.createObjectURL(file);
+  downloadFileByUrl(url, filename);
 };